style: исправить замечания по codestyle
